Rename top level controllers and views to be prefixed with Top to ensure they are easily identifiable and distinguishable.
