--- BudgetAnalyser/Dashboard/DashboardUserControl.xaml.cs
@@ -1,14 +1,0 @@
-﻿using System.Windows.Controls;
-
-namespace BudgetAnalyser.Dashboard;
-
-/// <summary>
-///     Interaction logic for DashboardUserControl.xaml
-/// </summary>
-public partial class DashboardUserControl : UserControl
-{
-    public DashboardUserControl()
-    {
-        InitializeComponent();
-    }
-}
