+++ BudgetAnalyser/Dashboard/TopDashboardUserControl.xaml.cs
@@ -1,0 +1,14 @@
+﻿using System.Windows.Controls;
+
+namespace BudgetAnalyser.Dashboard;
+
+/// <summary>
+///     Interaction logic for TopDashboardUserControl.xaml
+/// </summary>
+public partial class TopDashboardUserControl : UserControl
+{
+    public TopDashboardUserControl()
+    {
+        InitializeComponent();
+    }
+}
