--- BudgetAnalyser/Matching/MatchingRuleUserControl.xaml.cs
@@ -1,14 +1,0 @@
-﻿using System.Windows.Controls;
-
-namespace BudgetAnalyser.Matching;
-
-/// <summary>
-///     Interaction logic for MatchingRuleUserControl.xaml
-/// </summary>
-public partial class MatchingRuleUserControl : UserControl
-{
-    public MatchingRuleUserControl()
-    {
-        InitializeComponent();
-    }
-}
