+++ BudgetAnalyser/Matching/TopMatchingRuleUserControl.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="BudgetAnalyser.Matching.MatchingRuleUserControl"
+﻿<UserControl x:Class="BudgetAnalyser.Matching.TopMatchingRuleUserControl"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
