+++ BudgetAnalyser/Matching/TopMatchingRuleUserControl.xaml.cs
@@ -1,0 +1,14 @@
+﻿using System.Windows.Controls;
+
+namespace BudgetAnalyser.Matching;
+
+/// <summary>
+///     Interaction logic for TopMatchingRuleUserControl.xaml
+/// </summary>
+public partial class TopMatchingRuleUserControl : UserControl
+{
+    public TopMatchingRuleUserControl()
+    {
+        InitializeComponent();
+    }
+}
