+++ BudgetAnalyser/ReportsCatalog/TopReportsCatalogUserControl.xaml
@@ -1,0 +1,62 @@
+﻿<UserControl x:Class="BudgetAnalyser.ReportsCatalog.TopReportsCatalogUserControl"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:reportsCatalog="clr-namespace:BudgetAnalyser.ReportsCatalog"
+             mc:Ignorable="d"
+             d:DataContext="{d:DesignInstance reportsCatalog:TopReportsCatalogController}">
+
+    <Grid Background="{StaticResource Brush.TileBackground}">
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto" />
+            <RowDefinition Height="*" />
+        </Grid.RowDefinitions>
+        <WrapPanel Orientation="Horizontal">
+            <Border Style="{StaticResource ModernTileMediumStyle1}">
+                <Grid>
+                    <Rectangle Fill="{StaticResource Brush.Text.Default}"
+                               Height="30"
+                               HorizontalAlignment="Left"
+                               Margin="10"
+                               Width="30" />
+                    <Border BorderBrush="{StaticResource Brush.ModernTile.Background1}"
+                            BorderThickness="1,0,0,1"
+                            Height="24"
+                            HorizontalAlignment="Left"
+                            Margin="13,10,10,10"
+                            Width="24" />
+                    <Rectangle Fill="{StaticResource Brush.ModernTile.Background1}"
+                               Height="4"
+                               HorizontalAlignment="Left"
+                               Margin="14,-9,0,0"
+                               Width="4" />
+                    <Rectangle Fill="{StaticResource Brush.ModernTile.Background1}"
+                               Height="4"
+                               HorizontalAlignment="Left"
+                               Margin="14,0,0,0"
+                               Width="17" />
+                    <Rectangle Fill="{StaticResource Brush.ModernTile.Background1}"
+                               Height="4"
+                               HorizontalAlignment="Left"
+                               Margin="14,11,0,0"
+                               Width="13" />
+                    <Rectangle Fill="{StaticResource Brush.ModernTile.Background1}"
+                               Height="4"
+                               HorizontalAlignment="Left"
+                               Margin="14,20,0,0"
+                               Width="20" />
+                    <TextBlock Margin="50,10,10,10"
+                               Text="{Binding OverallPerformanceController.Title}"
+                               TextWrapping="Wrap" />
+                    <Button Click="OverallPerformanceClicked"
+                            IsEnabled="{Binding OverallPerformanceReportIsAvailable}"
+                            Style="{StaticResource Button.ModernTile.Overlay}" />
+                </Grid>
+            </Border>
+
+        </WrapPanel>
+        <Border Grid.Row="1" />
+    </Grid>
+
+</UserControl>
